refactor(api): extract security helpers
